+++ net/net.go
@@ -1,0 +1,2 @@
+// Package net provides models and utilities for GOBL Net, a decentralized identity and communication protocol.
+package net